Fix 3

+++ include/infinite-color-engine/InfiniteExponentialInterpolator.h
@@ -16,8 +16,8 @@ class InfiniteExponentialInterpolator : public InfiniteInterpolator
 public:
 	InfiniteExponentialInterpolator();
 
-	void setTargetColors(std::vector<linalg::aliases::float3>&& new_rgb_targets, float startTimeMs, bool debug = false) override;
-	void updateCurrentColors(float currentTimeMs, float minBrightness) override;
+	void setTargetColors(std::vector<linalg::aliases::float3>&& new_rgb_targets, long long startTimeMs, bool debug = false) override;
+	void updateCurrentColors(long long currentTimeMs, float minBrightness) override;
 	SharedOutputColors getCurrentColors(float minBrightness = 0.f) override;
 	void resetState() override;
 
@@ -31,7 +31,7 @@ private:
 
 	float _initialDuration = 150.0f;
 	float _tau  = 150.0f;
-	float _startAnimationTimeMs = 0.0f;
-	float _targetTime = 0.0f;
-	float _lastUpdate = 0.0f;	
+	long long _startAnimationTimeMs = 0;
+	long long _targetTime = 0;
+	long long _lastUpdate = 0;
 };
